chore: move `ImageProxyClient` to separate package
