+++ packages/image_proxy_client/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@apify/utilities",
+    "name": "@apify/image_proxy_client",
     "version": "1.0.0",
     "description": "Tools and constants shared across Apify projects.",
     "main": "dist/index.js",
@@ -43,8 +43,7 @@
         "access": "public"
     },
     "dependencies": {
-        "@apify/consts": "^1.0.0",
-        "@apify/log": "^1.0.0"
+        "create-hmac": "^1.1.7"
     },
     "devDependencies": {
         "@types/create-hmac": "^1.1.0"
