+++ packages/image_proxy_client/src/image_proxy_client.ts
