+++ packages/image_proxy_client/src/index.ts
@@ -1,0 +1,1 @@
+export * from './image_proxy_client';
