+++ packages/log/package.json
@@ -44,6 +44,6 @@
     },
     "dependencies": {
         "@apify/consts": "^1.0.0",
-        "chalk": "^4.0.0"
+        "ansi-colors": "^4.1.1"
     }
 }
