+++ packages/utilities/src/index.ts
@@ -2,7 +2,6 @@ export * from './utilities';
 export * from './utilities.client';
 export * from './exponential_backoff';
 export * from './health_checker';
-export * from './image_proxy_client';
 export * from './parse_jsonl_stream';
 export * from './streams_utilities';
 export * from './webhook_payload_template';
